refactor(src): :truck: rename DatePickerHelper to DateTimePickerHelper
